SANDOZ-Gen_AI_User_Training/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0

openai>=1.40.0
python-dotenv>=1.0.1,<2.0

pypdf>=4.2.0,<5.0

# Cliente Postgres (binario, evita problemas de compilación)
psycopg-c>=3.1.19,<4.0

# Adaptadores de pgvector para Python (no confundir con la extensión del servidor)
pgvector>=0.2.5,<0.3

pydantic>=2.6,<3.0
numpy>=1.26,<3.0



langchain
langchain-core
langchain-community
pymupdf
chromadb
langchain_text_splitters
scikit-learn
joblib
spacy